Refactor state-related functions into 'statehelpers'
As part of this, refactored some of the 'ValidateDirModeState' error handling to avoid a 'slice not updated' type of error (avoiding pointers to pointers)
